Move exercise hint to new course management routes
Split up hint components between participation and management
Fix lazy loading of hint components in code editor
Always load hint management components with course-management.module.ts
