+++ src/main/webapp/app/exercises/programming/manage/programming-exercise-detail.component.html
@@ -7,7 +7,12 @@
             <jhi-alert-error></jhi-alert-error>
             <div class="text-right">
                 <jhi-external-submission class="mr-1" [exercise]="programmingExercise"></jhi-external-submission>
-                <button type="submit" *ngIf="programmingExercise.isAtLeastInstructor" routerLink="/exercise/{{ programmingExercise.id }}/hints" class="btn btn-info btn-sm mr-1">
+                <button
+                    type="submit"
+                    *ngIf="programmingExercise.isAtLeastInstructor"
+                    [routerLink]="['/course-management', programmingExercise.course.id, 'exercises', programmingExercise.id, 'hints']"
+                    class="btn btn-info btn-sm mr-1"
+                >
                     <fa-icon [icon]="'pencil-alt'"></fa-icon>
                     <span class="d-none d-md-inline" jhiTranslate="entity.action.manageHints">Manage Hints</span>
                 </button>
