+++ src/main/webapp/app/exercises/shared/exercise-hint/manage/exercise-hint-detail.component.html
@@ -16,7 +16,7 @@
                 <dt><span jhiTranslate="artemisApp.exerciseHint.exercise">Exercise</span></dt>
                 <dd>
                     <div *ngIf="exerciseHint.exercise">
-                        <a [routerLink]="['/exercise', exerciseHint.exercise?.id, 'view']">{{ exerciseHint.exercise?.id }}</a>
+                        <a routerLink="../../../'view']">{{ exerciseHint.exercise?.id }}</a>
                     </div>
                 </dd>
             </dl>
@@ -25,7 +25,7 @@
                 <fa-icon [icon]="'arrow-left'"></fa-icon>&nbsp;<span jhiTranslate="entity.action.back"> Back</span>
             </button>
 
-            <button type="button" [routerLink]="['/exercise', exerciseId, 'hints', exerciseHint.id, 'edit']" class="btn btn-primary">
+            <button type="button" [routerLink]="['..', 'edit']" class="btn btn-primary">
                 <fa-icon [icon]="'pencil-alt'"></fa-icon>&nbsp;<span jhiTranslate="entity.action.edit"> Edit</span>
             </button>
         </div>
