+++ src/main/webapp/app/exercises/shared/exercise-hint/manage/exercise-hint.component.html
@@ -1,7 +1,7 @@
 <div>
     <h2 id="page-heading">
         <span jhiTranslate="artemisApp.exerciseHint.home.title">Exercise Hints</span>
-        <button id="jh-create-entity" class="btn btn-primary float-right jh-create-entity create-exercise-hint" routerLink="/exercise/{{ exerciseId }}/hints/new">
+        <button id="jh-create-entity" class="btn btn-primary float-right jh-create-entity create-exercise-hint" routerLink="new">
             <fa-icon [icon]="'plus'"></fa-icon>
             <span jhiTranslate="artemisApp.exerciseHint.home.createLabel">
                 Create new Exercise Hint
@@ -27,7 +27,7 @@
             <tbody>
                 <tr *ngFor="let exerciseHint of exerciseHints; trackBy: trackId">
                     <td>
-                        <a [routerLink]="['/exercise', exerciseId, 'hints', exerciseHint.id, 'view']">{{ exerciseHint.id }}</a>
+                        <a [routerLink]="[exerciseHint.id, 'view']">{{ exerciseHint.id }}</a>
                     </td>
                     <td>{{ exerciseHint.title }}</td>
                     <td [innerHTML]="exerciseHint.content | htmlForMarkdown"></td>
@@ -36,11 +36,11 @@
                     </td>
                     <td class="text-right">
                         <div class="btn-group">
-                            <button type="submit" [routerLink]="['/exercise', exerciseId, 'hints', exerciseHint.id, 'view']" class="btn btn-info btn-sm mr-1">
+                            <button type="submit" [routerLink]="[exerciseHint.id, 'view']" class="btn btn-info btn-sm mr-1">
                                 <fa-icon [icon]="'eye'"></fa-icon>
                                 <span class="d-none d-md-inline" jhiTranslate="entity.action.view">View</span>
                             </button>
-                            <button type="submit" [routerLink]="['/exercise', exerciseId, 'hints', exerciseHint.id, 'edit']" class="btn btn-primary btn-sm mr-1">
+                            <button type="submit" [routerLink]="[exerciseHint.id, 'edit']" class="btn btn-primary btn-sm mr-1">
                                 <fa-icon [icon]="'pencil-alt'"></fa-icon>
                                 <span class="d-none d-md-inline" jhiTranslate="entity.action.edit">Edit</span>
                             </button>
