+++ src/main/webapp/app/exercises/shared/exercise-hint/manage/exercise-hint.service.ts
