+++ src/main/webapp/app/exercises/shared/exercise-hint/participate/exercise-hint-participation.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { ArtemisSharedModule } from 'app/shared/shared.module';
+import { ExerciseHintStudentComponent, ExerciseHintStudentDialogComponent } from 'app/exercises/shared/exercise-hint/participate/exercise-hint-student-dialog.component';
+
+@NgModule({
+    imports: [ArtemisSharedModule],
+    declarations: [ExerciseHintStudentDialogComponent, ExerciseHintStudentComponent],
+    entryComponents: [ExerciseHintStudentDialogComponent],
+    exports: [ExerciseHintStudentDialogComponent, ExerciseHintStudentComponent],
+})
+export class ArtemisExerciseHintParticipationModule {}
