+++ src/main/webapp/app/exercises/shared/exercise-hint/participate/exercise-hint-student-dialog.component.ts
@@ -5,7 +5,7 @@ import { catchError, map, tap } from 'rxjs/operators';
 
 import { NgbActiveModal, NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 
-import { ExerciseHintService } from './exercise-hint.service';
+import { ExerciseHintService } from '../manage/exercise-hint.service';
 import { ExerciseHint } from 'app/entities/exercise-hint.model';
 
 /**
