+++ src/main/webapp/app/exercises/shared/exercise-hint/participate/exercise-hint-student-dialog.scss
