+++ src/test/javascript/spec/mocks/mock-exercise-hint.service.ts
@@ -1,7 +1,7 @@
 import { HttpResponse } from '@angular/common/http';
 import { Observable, of } from 'rxjs';
 import { ExerciseHint } from 'app/entities/exercise-hint.model';
-import { ExerciseHintResponse, IExerciseHintService } from 'app/exercises/shared/exercise-hint/exercise-hint.service';
+import { ExerciseHintResponse, IExerciseHintService } from 'app/exercises/shared/exercise-hint/manage/exercise-hint.service';
 
 export class MockExerciseHintService implements IExerciseHintService {
     private exerciseHintDummy = { id: 1 } as ExerciseHint;
